fsm ferdig

+++ assorted_funcs.h
@@ -20,7 +20,7 @@ Starte the motor.
 @param current_floor	Elevators current floor
 @param desired_floor	The requested floor to go to
 */
-void move_to_floor(int current_floor, int desired_floor);
+void move_to_floor(double current_floor, int desired_floor);
 
 
 /*
